move PosOfWindow into its own package
